fix: emmcCommonUpdateProcess
